Move aggregators package to under openwifirrm
