+++ src/test/java/com/facebook/openwifirrm/modules/ModelerUtilsTest.java
@@ -20,8 +20,8 @@ import java.util.LinkedList;
 import org.junit.jupiter.api.Test;
 
 import com.facebook.openwifirrm.DeviceDataManager;
+import com.facebook.openwifirrm.aggregators.MeanAggregator;
 import com.facebook.openwifirrm.modules.Modeler.DataModel;
-import com.facebook.openwifirrm.modules.aggregators.MeanAggregator;
 import com.facebook.openwifirrm.optimizers.TestUtils;
 import com.facebook.openwifirrm.ucentral.UCentralUtils.WifiScanEntry;
