+++ src/test/java/com/facebook/openwifirrm/modules/aggregators/MeanAggregatorTest.java
@@ -12,6 +12,8 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.Test;
 
+import com.facebook.openwifirrm.aggregators.MeanAggregator;
+
 public class MeanAggregatorTest {
 	@Test
 	public void testEmptyAndNonEmptyAndReset() {
